+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `acquire-device-properties`: A `uint8_t` member `enable_multiscale` of `StorageProperties`.
+- `acquire-device-properties`: A corresponding entry in `StoragePropertyMetadata`.
+- `acquire-device-properties`: A convenience function for enabling multiscale, similar to the function for
+  setting chunking properties.
+
 ## [0.1.3](https://github.com/acquire-project/acquire-core-libs/compare/v0.1.2...v0.1.3) - 2023-06-27
 
 ### Changed